move tests to katgpucbf
