+++ test/qualification/vlbi_decoder/test_vtp_decoder.py
@@ -1,11 +1,9 @@
-#! /usr/bin/env python3
-
 ################################################################################
 # Copyright (c) 2026, National Research Foundation (SARAO)
 #
 # Licensed under the BSD 3-Clause License (the "License"); you may not use
-# this file except in compliance with the License. You may obtain a copy
-# of the License at
+# this file except in compliance with the License. You may obtain a copy of
+# the License at
 #
 #   https://opensource.org/licenses/BSD-3-Clause
 #
@@ -16,28 +14,17 @@
 # limitations under the License.
 ################################################################################
 
-"""Check for :class:`qualification.recv.VTPDecoder`.
+"""Unit tests for :mod:`qualification.vlbi_decoder.vtp_decoder`."""
 
-Run directly:
-    python qualification/test/test_vtp_decoder.py
-"""
-
-import asyncio
 import io
 import struct
-import sys
-from pathlib import Path
-
-sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
 import numpy as np
 from baseband.vdif import VDIFFrame
 
-from qualification.recv import VTPBuffer, VTPDecoder
+from qualification.vlbi_decoder.vtp_decoder import VTPBuffer, VTPDecoder
 
 SAMPLES_PER_FRAME = 8000
-BANDWIDTH = 64000000
-N_THREADS = 4
 
 
 def make_vtp_packet(
@@ -60,18 +47,18 @@ def make_vtp_packet(
     return struct.pack("<Q", seq_id) + buf.getvalue()
 
 
-def check_samples_per_frame_from_first_frame() -> None:
-    """`check_samples_per_frame_from_first_frame` checks that the first frame sets the samples_per_frame."""
+def test_samples_per_frame_from_first_frame() -> None:
+    """First frame sets samples_per_frame on the buffer."""
     buffer = VTPBuffer()
     buffer.add_packet(make_vtp_packet(0, frame_nr=1, seconds=100, thread_id=0))
     assert buffer.samples_per_frame == SAMPLES_PER_FRAME
     assert buffer.seq_ids == [0]
 
 
-async def check_decode_vdif_framesets_filters_incomplete_threads() -> None:
-    """`check_decode_vdif_framesets_filters_incomplete_threads` checks that decode_vdif_framesets correctly filters out framesets with incomplete threads."""  # noqa: E501
+def test_decode_vdif_framesets_filters_incomplete_threads() -> None:
+    """Incomplete thread sets are recorded as invalid framesets."""
     buffer = VTPBuffer()
-    # 2 theads data only
+    # 2 threads of data only
     buffer.add_packet(make_vtp_packet(0, frame_nr=0, seconds=100, thread_id=0))
     buffer.add_packet(make_vtp_packet(1, frame_nr=0, seconds=100, thread_id=1))
     # 4 threads but only 3 unique thread ids
@@ -81,13 +68,13 @@ async def check_decode_vdif_framesets_filters_incomplete_threads() -> None:
     buffer.add_packet(make_vtp_packet(5, frame_nr=1, seconds=100, thread_id=0))
     decoder = VTPDecoder(buffer, 4)
 
-    framesets = [item for item in decoder.vtp_framesets()]
+    framesets = list(decoder.vtp_framesets())
     assert len(framesets) == 0
-    assert len(decoder.invalid_framesets) == 2, f"Invalid framesets: {decoder.invalid_framesets}"
+    assert len(decoder.invalid_framesets) == 2
 
 
-async def check_decode_vdif_framesets_filters_duplicate_seq_ids() -> None:
-    """`check_decode_vdif_framesets_filters_duplicate_seq_ids` checks that decode_vdif_framesets correctly filters out framesets with duplicate sequence IDs."""  # noqa: E501
+def test_decode_vdif_framesets_filters_duplicate_seq_ids() -> None:
+    """Framesets with duplicate sequence IDs are filtered out."""
     buffer = VTPBuffer()
     buffer.add_packet(make_vtp_packet(4, frame_nr=1, seconds=100, thread_id=0))
     buffer.add_packet(make_vtp_packet(5, frame_nr=1, seconds=100, thread_id=1))
@@ -96,13 +83,13 @@ async def check_decode_vdif_framesets_filters_duplicate_seq_ids() -> None:
     buffer.add_packet(make_vtp_packet(5, frame_nr=3, seconds=100, thread_id=1))
     decoder = VTPDecoder(buffer, 2)
 
-    framesets = [item for item in decoder.vtp_framesets()]
+    framesets = list(decoder.vtp_framesets())
     assert len(framesets) == 1
     assert len(decoder.invalid_framesets) == 2
 
 
-async def check_decode_vdif_framesets_unordered() -> None:
-    """`check_decode_vdif_framesets_unordered` checks that decode_vdif_framesets yields on each frameset boundary."""
+def test_decode_vdif_framesets_unordered() -> None:
+    """Framesets are yielded correctly when packets arrive out of order."""
     buffer = VTPBuffer()
     buffer.add_packet(make_vtp_packet(5, frame_nr=10, seconds=99, thread_id=0))
     buffer.add_packet(make_vtp_packet(6, frame_nr=0, seconds=100, thread_id=0))
@@ -114,17 +101,16 @@ async def check_decode_vdif_framesets_unordered() -> None:
     buffer.add_packet(make_vtp_packet(10, frame_nr=0, seconds=100, thread_id=3))
     buffer.add_packet(make_vtp_packet(12, frame_nr=1, seconds=100, thread_id=2))
     decoder = VTPDecoder(buffer, 4)
-    # todo: just do random access in a test instead.
 
-    framesets = [item for item in decoder.vtp_framesets()]
-    assert len(decoder.invalid_framesets) == 1, f"Invalid framesets: {decoder.invalid_framesets}"
+    framesets = list(decoder.vtp_framesets())
+    assert len(decoder.invalid_framesets) == 1
     assert len(framesets) == 2
     assert framesets[0] == ([6, 8, 9, 10], (100, 0))
     assert framesets[1] == ([7, 11, 12, 13], (100, 1))
 
 
-async def check_decode_vdif_framesets_second_border() -> None:
-    """`check_decode_vdif_framesets_second_border` checks that decode_vdif_framesets yields on each frameset boundary over multiple seconds."""  # noqa: E501
+def test_decode_vdif_framesets_second_border() -> None:
+    """Framesets are grouped correctly across a second boundary."""
     buffer = VTPBuffer()
     buffer.add_packet(make_vtp_packet(0, frame_nr=0, seconds=100, thread_id=0))
     buffer.add_packet(make_vtp_packet(1, frame_nr=0, seconds=100, thread_id=1))
@@ -135,9 +121,8 @@ async def check_decode_vdif_framesets_second_border() -> None:
     buffer.add_packet(make_vtp_packet(6, frame_nr=1, seconds=101, thread_id=0))
     buffer.add_packet(make_vtp_packet(7, frame_nr=1, seconds=101, thread_id=1))
     decoder = VTPDecoder(buffer, 2)
-    # todo: just do random access in a test instead.
 
-    framesets = [item for item in decoder.vtp_framesets()]
+    framesets = list(decoder.vtp_framesets())
     assert len(decoder.invalid_framesets) == 0
     assert len(framesets) == 4
     assert framesets[0] == ([0, 1], (100, 0))
@@ -146,8 +131,8 @@ async def check_decode_vdif_framesets_second_border() -> None:
     assert framesets[3] == ([6, 7], (101, 1))
 
 
-def check_close_clears_state() -> None:
-    """`check_close_clears_state` checks that clear() resets all buffered state."""
+def test_clear_clears_state() -> None:
+    """clear() resets all buffered state."""
     buffer = VTPBuffer()
     buffer.add_packet(make_vtp_packet(0, frame_nr=0, seconds=100, thread_id=0))
     buffer.add_packet(make_vtp_packet(1, frame_nr=0, seconds=100, thread_id=0))
@@ -157,34 +142,3 @@ def check_close_clears_state() -> None:
     assert buffer.thread_ids == []
     assert buffer.samples_per_frame is None
     assert buffer.frame_ids == []
-
-
-def main() -> None:
-    """Run all checks."""
-    checks = [check_close_clears_state, check_samples_per_frame_from_first_frame]
-    for check in checks:
-        print(f"running {check.__name__}...", flush=True)
-        check()
-        print("  ok", flush=True)
-
-    print("running check_decode_vdif_framesets_filters_incomplete_threads...", flush=True)
-    asyncio.run(check_decode_vdif_framesets_filters_incomplete_threads())
-    print("  ok", flush=True)
-
-    print("running check_decode_vdif_framesets_filters_duplicate_seq_ids...", flush=True)
-    asyncio.run(check_decode_vdif_framesets_filters_duplicate_seq_ids())
-    print("  ok", flush=True)
-
-    print("running check_decode_vdif_framesets_unordered...", flush=True)
-    asyncio.run(check_decode_vdif_framesets_unordered())
-    print("  ok", flush=True)
-
-    print("running check_decode_vdif_framesets_second_border...", flush=True)
-    asyncio.run(check_decode_vdif_framesets_second_border())
-    print("  ok", flush=True)
-
-    print("all checks passed")
-
-
-if __name__ == "__main__":
-    main()
